Require three-way handshake before creating flow
Require the TCP connection to fully establish, prior to creating the NAT rule.

TCP simultaneous open should be supported as well, by ensuring SEQs are
recorded when seen by each party, and not expecting an ACK to arive alongside
a SEQ in the case of the connectee's portion of the handshake.