Use FlowRemoved events to timeout NAT entries
Once a connection establishes itself, we'll rely on our switch to correctly
timeout the connection, since the controller won't be seeing packets anymore.
When we see that a flow has timed out, we should cancel the other side of the
flow, and remove its entry from the NAT table.  Additionally, if this was the
last NAT table entry using the given NAT port, we'll recycle that NAT port.